requirements.txt
altgraph==0.17.4
asgiref==3.9.1
av==14.4.0
certifi==2025.7.14
charset-normalizer==3.4.2
customtkinter==5.2.2
darkdetect==0.8.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pefile==2023.2.7
psycopg2-binary==2.9.10
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.8
PyJWT==2.10.1
python-decouple==3.8
pywin32-ctypes==0.2.3
requests==2.32.4
setuptools==80.9.0
sqlparse==0.5.3
tkinterdnd2==0.4.3
tkinterdnd2-universal==1.7.3
tzdata==2025.2
urllib3==2.5.0
